Getting the Swagger UI React component working
I had to downgrade to React 18 to get around some dependency issues with swagger-react.

Added a dependency to react-router so I can do actual routing.

Due to the spec being hosted from the Spring server, Vite will not be able to display Swagger correctly.
Not that Swagger via Vite is an intended usecase.